Automatic batching of fcwt calculations on long sequences

Performs a fast continuous wavelet transform on long sequences. It should not be used on short sequences since boundary artefacts are discarded (and those potentially dominate for short sequences.)

Usage

fcwt_batch(
  signal,
  sample_freq,
  freq_begin,
  freq_end,
  n_freqs,
  sigma,
  max_batch_size = ceiling(4 * 10^9/(n_freqs * 4)),
  time_resolution,
  nthreads = 8L
)

Arguments

signal

Real-valued time series. The time step gaps are assumed to be evenly spaced.

sample_freq

Sampling rate of input time series in physical units.

freq_begin

Sets the minimal frequency (> 0), in same units as sample_freq, that should be contained in the output.

freq_end

Sets the maximal frequency (> freq_begin), in same units as sample_freq, that should be contained in the output.

n_freqs

Number of frequency bins generated by the CWT. The frequencies are linearly distributed. Computation time increases when raising the number of frequency bins.

sigma

Sets a dimensionless parameter modifying the wavelet spread which in the time-domain is roughly given by \(\Sigma_t \sim \sqrt{2} \frac{\sigma}{f}\). Changing this parameter adjusts the time/frequency uncertainty balance. Defaults to 1. Larger (lower) value of sigma corresponds to a better (worse) frequency resolution and a worse (better) time resolution.

max_batch_size

The maximal batch size that is used for splitting up the input sequence. This limits the maximal memory that is used. Defaults to roughly 4GB. The actual batch size is optimized for use with FFTW.

time_resolution

The time resolution in inverse units of sample_freq of the result. Memory consumption is directly related to that. Can not be higher than the time resolution of the input signal.

nthreads

Number of threads used by the computation, if applicable.

Details

In case of input sequences that exceed the a certain size, the output sequence will not fit into the local memory and the fcwt cannot be performed in one run. For instance, in case of processing a song of 10 minutes length (assuming a sampling rate of 44100 Hz), the size of the output vector is 10 * 60 seconds * 44100 Hz * nfreqs * 4 bytes, which for e.g. nfreqs = 200, equals ~ 21 GB, hence nowadays already at the limit of the hardware of a modern personal computer.

In cases where the needed output time-resolution is smaller than the time resolution of the input signal, one can perform the fcwt and reduce the output size by averaging. (The input signal time resolution can in general not be reduced since high-frequency information would get lost.)

This function splits up the input sequence into batches, processes each batch separately, reduces the time resolution, and adds the outputs together.

Attention: Boundary artefacts are removed, so some high frequency information at the beginning and the end of the sequence is lost. (The amount depends on the minimal frequency captured min_freq.)

See also

Examples

res <-
  fcwt_batch(
    ts_sin_sin,
    sample_freq = 44100,
    freq_begin = 100,
    freq_end = 11000,
    n_freqs = 300,
    sigma = 10,
    max_batch_size = 20000,
    time_resolution = 0.001
  )
#>   |                                                                              |                                                                      |   0%  |                                                                              |======                                                                |   9%  |                                                                              |=============                                                         |  18%  |                                                                              |===================                                                   |  27%  |                                                                              |=========================                                             |  36%  |                                                                              |===============================                                       |  45%  |                                                                              |======================================                                |  54%  |                                                                              |============================================                          |  63%  |                                                                              |==================================================                    |  72%  |                                                                              |======================================================================| 100%